add getUserInfo
